+++ ansibleplaybookgrapher/__init__.py
@@ -1,2 +1,4 @@
+from .parser import PlaybookParser
+
 __version__ = "0.12.0-dev"
 __prog__ = "ansible-playbook-grapher"
